requirements.txt
-e .
torch
pytorch-metric-learning
faiss-cpu
pyyaml
argparse